requirements.txt
alembic==1.5.5
aniso8601==9.0.1
APScheduler==3.7.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
